Add minutes and second output
The format in which time is being outputted on the information board
is as follows: m : s ... getTime() function returns minutes, seconds or
nettime based on what character is passed in.

e.g object.getTime('m') returns minutes
    object.getTime('s') returns seconds
    object.getTime('n') returns net time

I changed the return data type from double to integer, feel free to edit
this back to double if needed.